Move ApplyTransaction to blockchain.go
